Move helpers to proper namespace instead of class
